+++ util/src/main/java/dk/itu/util/shape/WayPath.java
@@ -11,12 +11,12 @@ import static dk.itu.util.PolygonUtils.isClosed;
 public class WayPath implements Shape {
     private static final AffineTransform transform = new AffineTransform();
     private final byte[] pointTypes;
-    private final float[] outerCoordinates;
+    private final double[] outerCoordinates;
     private final PathIterator pathIteratorNodeSkip;
 
     private short pathIteratorPointer = 0;
 
-    public WayPath(float[] coordinates) {
+    public WayPath(double[] coordinates) {
         this.outerCoordinates = coordinates;
         // Set point types
         this.pointTypes = new byte[coordinates.length / 2];
@@ -80,7 +80,7 @@ public class WayPath implements Shape {
         };
     }
 
-    public float[] getOuterCoordinates() {
+    public double[] getOuterCoordinates() {
         return outerCoordinates;
     }
